services/desktop_server/requirements.txt
fastapi
fastapi_utilities
httpx
ipywidgets
jinja2
jupyterlab
jupytext
matplotlib
mypy
numba
numpy
pandas
pillow
pip
pre-commit
pydantic
pydantic_settings
pytest
python-multipart
rich
scipy
setuptools
sqlmodel
tdqm
torch
uvicorn

# diffusion
accelerate
diffusers
transformers
huggingface_hub
torch
torchaudio
torchvision

# Flux dep
protobuf
sentencepiece
